# docs/core/CORE_API_OPENAPI.yaml
openapi: 3.1.1
info:
  title: Вузы России — API ядра-бекенда
  version: 0.1.0
  description: 'Предложенный контракт MVP от 2026-09-24. Не описание текущей реализации. Только операции ядра, включая
    API, вызываемое backend сервисов. Доменные API других компонентов не входят. CORE_API_SPEC.md дополняет HTTP-контракт
    правилами JWT, изоляции, транзакций и bootstrap. Private операции имеют отдельный server и AND-требование machine
    + actor; на публичном ingress всегда 404. x-required-permission, x-required-machine-scope и x-network-access
    — документационные extensions, не автоматическое обеспечение доступа. Форматы JSON: snake_case, UUID string,
    UTC. Формат документа — OpenAPI 3.1.1 (https://spec.openapis.org/oas/v3.1.1.html).'
servers:
- url: https://core.example
  description: Публичный HTTPS origin ядра; заменить при развёртывании.
security:
- CoreBearer: []
...
paths:
...
  /api/v1/internal/service/{service_id}/manifest:
    parameters:
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Service registry
      summary: Прочитать зарегистрированный manifest своего экземпляра
      operationId: getOwnManifest
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceManifest'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - MachineBearer: []
      x-required-machine-scope: manifest:write
      description: ' Требуемый machine scope: manifest:write.'
    put:
      tags:
      - Service registry
      summary: Заменить имя и меню своего экземпляра
      operationId: replaceOwnManifest
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceManifest'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Не меняет endpoints, tenant, тип, профили или enabled. ID меню уникальны. Разрешено для disabled
        instance при active вузе. Требуемый machine scope: manifest:write.'
      security:
      - MachineBearer: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceManifest'
      x-required-machine-scope: manifest:write
  /api/v1/internal/service/{service_id}/users/{user_id}/profiles:
    parameters:
    - $ref: '#/components/parameters/service_id'
    - $ref: '#/components/parameters/user_id'
    get:
      tags:
      - Service RBAC
      summary: Прочитать профили участника в вузе своего экземпляра
      operationId: getServiceUserProfiles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Только чтение; target пользователь должен состоять в вузе экземпляра. Нет write-операций для
        профилей в публичном internal API. Требуемый machine scope: profiles:read.'
      security:
      - MachineBearer: []
      x-required-machine-scope: profiles:read
  /api/v1/internal/service/{service_id}/roles:
    parameters:
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Service RBAC
      summary: Получить определения ролей своего экземпляра
      operationId: listOwnRoles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - MachineBearer: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      x-required-machine-scope: roles:read
      description: ' Требуемый machine scope: roles:read.'
    post:
      tags:
      - Service RBAC
      summary: Создать роль своего экземпляра
      operationId: createOwnRole
      responses:
        '201':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Код уникален и неизменяем; permissions из словаря типа. Administration role definitions защищены,
        любые записи через этот маршрут — 403. Требуемый machine scope: roles:write.'
      security:
      - MachineBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      x-required-machine-scope: roles:write
  /api/v1/internal/service/{service_id}/roles/{role_code}:
    parameters:
    - $ref: '#/components/parameters/service_id'
    - $ref: '#/components/parameters/role_code'
    get:
      tags:
      - Service RBAC
      summary: Получить определение роли
      operationId: getOwnRole
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - MachineBearer: []
      x-required-machine-scope: roles:read
      description: ' Требуемый machine scope: roles:read.'
    patch:
      tags:
      - Service RBAC
      summary: Изменить имя или права роли
      operationId: updateOwnRole
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Системные роли administration неизменяемы. Сужение allowed_profiles при несовместимых назначениях
        — 409 ROLE_IN_USE. Требуемый machine scope: roles:write.'
      security:
      - MachineBearer: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePatch'
      x-required-machine-scope: roles:write
    delete:
      tags:
      - Service RBAC
      summary: Удалить неназначенную роль
      operationId: deleteOwnRole
      responses:
        '204':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Есть назначения — 409 ROLE_IN_USE. Системные роли administration защищены. Неизвестная роль
        — 404. Требуемый machine scope: roles:write.'
      security:
      - MachineBearer: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      x-required-machine-scope: roles:write
  /api/v1/internal/service/{service_id}/users/{user_id}/profiles/{profile}/roles:
    parameters:
    - $ref: '#/components/parameters/service_id'
    - $ref: '#/components/parameters/user_id'
    - $ref: '#/components/parameters/profile_path'
    get:
      tags:
      - Service RBAC
      summary: Прочитать назначенные роли и эффективные permissions
      operationId: getOwnAssignments
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentSet'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - MachineBearer: []
      x-required-machine-scope: assignments:read
      description: ' Требуемый machine scope: assignments:read.'
    put:
      tags:
      - Service RBAC
      summary: Атомарно заменить назначения ролей в своём экземпляре
      operationId: replaceOwnAssignments
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentSet'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Участник и профиль должны существовать. [] снимает роли. Нет доступа к другому экземпляру. Для
        administration любые изменения этим маршрутом — 403. Требуемый machine scope: assignments:write.'
      security:
      - MachineBearer: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentInput'
      x-required-machine-scope: assignments:write
...
  /api/v1/institution/{institution_id}/internal/services/{service_id}/roles:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Private roles
      summary: Прочитать роли выбранного экземпляра вуза
      operationId: listInstitutionServiceRoles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: services.read.'
      x-required-permission: services.read
  /api/v1/institution/{institution_id}/internal/services/{service_id}/users/{user_id}/profiles/{profile}/roles:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/service_id'
    - $ref: '#/components/parameters/user_id'
    - $ref: '#/components/parameters/profile_path'
    get:
      tags:
      - Private roles
      summary: Прочитать назначения участника в экземпляре вуза
      operationId: getInstitutionServiceAssignments
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentSet'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: roles.manage.'
      x-required-permission: roles.manage
    put:
      tags:
      - Private roles
      summary: Назначить или снять роли участника в экземпляре вуза
      operationId: replaceInstitutionServiceAssignments
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentSet'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Для administration дополнительно требуется
        текущая роль owner; последний owner защищён. Роли должны существовать именно в этом экземпляре и допускать
        профиль. Требуемый permission actor: roles.manage.'
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentInput'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: roles.manage
...
  /api/v1/internal/service/{service_id}/groups:
    parameters:
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Service RBAC
      summary: Учебные группы вуза своего экземпляра
      operationId: listServiceGroups
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Только чтение.'
      x-extension: study-groups
      security:
      - MachineBearer: []
      x-required-machine-scope: groups:read
  /api/v1/internal/service/{service_id}/groups/{group_id}/members:
    parameters:
    - $ref: '#/components/parameters/service_id'
    - $ref: '#/components/parameters/group_id'
    get:
      tags:
      - Service RBAC
      summary: Состав учебной группы
      operationId: getServiceGroupMembers
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupMembers'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Только чтение.'
      x-extension: study-groups
      security:
      - MachineBearer: []
      x-required-machine-scope: groups:read
  /api/v1/internal/service/{service_id}/members:
    parameters:
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Service RBAC
      summary: Участники вуза своего экземпляра
      operationId: listServiceMembers
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMemberList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Профили и имя из регистрации (расширение для сервиса «Люди»).'
      x-extension: study-groups
      security:
      - MachineBearer: []
      x-required-machine-scope: profiles:read
...
components:
  securitySchemes:
    CoreBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Пользовательский token_use=core_access, aud=core-api; живая sid обязательна.
    MachineBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token_use=machine_access, aud=core-internal. Credential и tenant проверяются на каждом запросе.
        Не пользовательский токен.
    ServiceBasic:
      type: http
      scheme: basic
      description: 'Только HTTPS token exchange: client_id как username, client_secret как password.'
    ActorToken:
      type: apiKey
      in: header
      name: X-Actor-Token
      description: 'JWT без префикса Bearer: service_access человека для экземпляра administration, профиль admin.
        AND с MachineBearer; актуальный RBAC из БД.'
  parameters:
...
    service_id:
      name: service_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  headers:
    RequestId:
      description: UUID запроса, назначенный ядром.
      schema:
        $ref: '#/components/schemas/UUID'
    NoStore:
      description: Ответ нельзя сохранять в HTTP-кэше.
      schema:
        type: string
        const: no-store
    ETag:
      description: Strong ETag ресурса для If-Match.
      schema:
        type: string
        examples:
        - '"rev-7"'
...